+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.1.0
 
 * Add sounds for bottling & using experience
+* Add override for dropped xp amount on death
 
 # 2.0.0